refactor(*): break up into individual modules
this should make it much easier to comprehend
and write tests for karma-webpack.
there is one change in how the
KarmaWebpackController is managed, we now
instantiate this in the preprocessor phase
and propagate the value within the karma config
object as a private variable. This allows
for breaking the framework and preprocessor
into separates modules and has the added
benefit of being able to run multiple
times in a given session without sharing
mutable state. This allows integrations
tests to be run in parallel as well as multiple
times which was previously not possible.

Fixes N/A
