+++ lib/index.js
@@ -1,1 +1,1 @@
-module.exports = require('./karma-webpack');
+module.exports = require('./karma/plugin');
