+++ lib/karma-webpack/preprocessor.js
@@ -1,45 +1,12 @@
 const path = require('path');
-const fs = require('fs');
 
 const glob = require('glob');
 const minimatch = require('minimatch');
 
-const { ensureWebpackFrameworkSet } = require('./karma/karmaConfigValidator');
+const { ensureWebpackFrameworkSet } = require('../karma/validation');
+const { hash } = require('../utils/hash');
 
-const { hash } = require('./utils/hash');
-
-const { KarmaWebpackController } = require('./KarmaWebpackController');
-
-const controller = new KarmaWebpackController();
-
-function registerExtraWebpackFiles(config, _controller) {
-  const localController = _controller || controller;
-  const commonsPath = path.join(localController.outputPath, 'commons.js');
-  const runtimePath = path.join(localController.outputPath, 'runtime.js');
-
-  // make sure tmp folder exists
-  if (!fs.existsSync(localController.outputPath)) {
-    fs.mkdirSync(localController.outputPath);
-  }
-
-  // create dummy files for commons.js and runtime.js so they get included by karma
-  fs.closeSync(fs.openSync(commonsPath, 'w'));
-  fs.closeSync(fs.openSync(runtimePath, 'w'));
-
-  // register for karma
-  config.files.unshift({
-    pattern: commonsPath,
-    included: true,
-    served: true,
-    watched: false,
-  });
-  config.files.unshift({
-    pattern: runtimePath,
-    included: true,
-    served: true,
-    watched: false,
-  });
-}
+const KW_Controller = require('./controller');
 
 function getPathKey(filePath, withExtension = false) {
   const pathParts = path.parse(filePath);
@@ -81,7 +48,9 @@ function configToWebpackEntries(config) {
   return webpackEntries;
 }
 
-function preprocessorFactory(config, emitter) {
+function KW_Preprocessor(config, emitter) {
+  const controller = new KW_Controller();
+  config.__karmaWebpackController = controller;
   ensureWebpackFrameworkSet(config);
 
   // one time setup
@@ -118,12 +87,6 @@ function preprocessorFactory(config, emitter) {
   };
 }
 
-registerExtraWebpackFiles.$inject = ['config'];
-preprocessorFactory.$inject = ['config', 'emitter'];
+KW_Preprocessor.$inject = ['config', 'emitter'];
 
-module.exports = {
-  'preprocessor:webpack': ['factory', preprocessorFactory],
-  'framework:webpack': ['factory', registerExtraWebpackFiles],
-  registerExtraWebpackFiles,
-  configToWebpackEntries,
-};
+module.exports = KW_Preprocessor;
