+++ lib/karma/validation.js
