+++ test/unit/karma-webpack/preprocessor.test.js
@@ -1,0 +1,8 @@
+const KW_Preprocessor = require('../../../lib/karma-webpack/preprocessor');
+
+describe('KW_Preprocessor', () => {
+  it('should be defined', () => {
+    expect(KW_Preprocessor).toBeDefined();
+  });
+  // todo(mikol): write a KW_Preprocessor test suite before v5 official release.
+});
