+++ test/unit/karma/plugin.test.js
@@ -1,0 +1,8 @@
+const KW_KarmaPlugin = require('../../../lib/karma/plugin');
+
+describe('KW_KarmaPlugin', () => {
+  it('should be defined', () => {
+    expect(KW_KarmaPlugin).toBeDefined();
+  });
+  // todo(mikol): write test suite for KarmaWebpackPlugin prior to v5 official release.
+});
