+++ test/unit/webpack/defaults.test.js
@@ -1,0 +1,8 @@
+const DefaultWebpackOptionsFactory = require('../../../lib/webpack/defaults');
+
+describe('DefaultWebpackOptionsFactory', () => {
+  it('should be defined', () => {
+    expect(DefaultWebpackOptionsFactory).toBeDefined();
+    // todo(mikol): write a DefaultWebpackOptionsFactory test suite before v5 official release.
+  });
+});
