+++ test/unit/webpack/plugin.test.js
@@ -1,0 +1,8 @@
+const KW_WebpackPlugin = require('../../../lib/webpack/plugin');
+
+describe('KW_WebpackPlugin', () => {
+  it('should be defined', () => {
+    expect(KW_WebpackPlugin).toBeDefined();
+    // todo(mikol): write a KW_WebpackPlugin test suite before v5 official release.
+  });
+});
